Enum ethcore_miner::gas_pricer::GasPricer

pub enum GasPricer {
    Fixed(U256),
    Calibrated(GasPriceCalibrator),
}

Struct to look after updating the acceptable gas price of a miner.

Variants
Fixed(U256)

A fixed gas price in terms of Wei - always the argument given.

Calibrated(GasPriceCalibrator)

Gas price is calibrated according to a fixed amount of USD.

Methods

impl GasPricer

pub fn new_calibrated(
options: GasPriceCalibratorOptions,
fetch: FetchClient,
p: CpuPool
) -> GasPricer

Create a new Calibrated GasPricer.

pub fn new_fixed(gas_price: U256) -> GasPricer

Create a new Fixed GasPricer.

pub fn recalibrate<F: FnOnce(U256) + Sync + Send + 'static>(
&mut self,
set_price: F
)

Recalibrate current gas price.