Struct ethcore_miner::pool::local_transactions::LocalTransactionsList

pub struct LocalTransactionsList { /* fields omitted */ }

Keeps track of local transactions that are in the queue or were mined/dropped recently.

Methods

impl LocalTransactionsList

pub fn new(max_old: usize) -> Self

Create a new list of local transactions.

pub fn contains(&self, hash: &H256) -> bool

Returns true if the transaction is already in local transactions.

pub fn all_transactions(&self) -> &LinkedHashMap<H256, Status>

Return a map of all currently stored transactions.

pub fn has_pending(&self) -> bool

Returns true if there are pending local transactions.