Struct ethcore_miner::pool::VerifiedTransaction

pub struct VerifiedTransaction { /* fields omitted */ }

Verified transaction stored in the pool.

Methods

impl VerifiedTransaction

pub fn from_pending_block_transaction(tx: SignedTransaction) -> Self

Create VerifiedTransaction directly from SignedTransaction.

This method should be used only:

for tests

In case we are converting pending block transactions that are already in the queue to match the function signature.

pub fn signed(&self) -> &SignedTransaction

Gets wrapped SignedTransaction

pub fn pending(&self) -> &PendingTransaction

Gets wrapped PendingTransaction