Enum ethcore_miner::pool::verifier::Transaction

pub enum Transaction {
    Unverified(UnverifiedTransaction),
    Retracted(UnverifiedTransaction),
    Local(PendingTransaction),
}

Transaction to verify.

Variants
Unverified(UnverifiedTransaction)

Fresh, never verified transaction.

We need to do full verification of such transactions

Retracted(UnverifiedTransaction)

Transaction from retracted block.

We could skip some parts of verification of such transactions

Local(PendingTransaction)

Locally signed or retracted transaction.

We can skip consistency verifications and just verify readiness.

Methods

impl Transaction

pub fn hash(&self) -> H256

Return transaction hash

pub fn gas_price(&self) -> &U256

Return transaction gas price